pyproject.toml
[tool.poetry]
name = "hierarchical_data"
version = "0.1.0"
description = "A simple Django server that uses MPTT models from django-mptt to create a Dropbox-esque web interface where you can create \"folders\" and \"files\" in an arbitrary structure and then display that structure."
authors = ["Reginald Tjitradi <rtjitradi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
